Implement find-my-phone
Initially for Sailfish only and GTS and above devices.  Listen for event from watch, then
use pulse-audio to play a sound effect, which bypasses any "silent" mode
on Sailfish.

Listen for the cencel event and stop playing the sound,

Sound is played using a 5 second timer, for a maximum of 10 times.