Refactor how operations are handled
Creates a new abstracct QBLEService that has the concept of operations.
This is the bare minimum refactor in preparation for further refatoing,
and the addition of some more operations which support the amazfit
balance, starting with an 'init' operation.  The intention is to not
have to hard code all the operations into the services.